fix(code): address PR #150 review (cross-arch key, fmt, test retry)
CodeRabbit triage on #150:

- [fix] keys.rs: hash the line as a fixed-width u64, not usize::to_le_bytes().
  usize is 8 bytes on 64-bit and 4 on 32-bit, so the same symbol could key
  differently across architectures. On 64-bit the bytes are unchanged, so no
  re-keying; 32-bit now agrees.
- [fix] ra_span_agreement.rs: poll documentSymbol with a short backoff instead
  of failing on the first empty result (rust-analyzer can settle briefly after
  start).
- [fix] cargo fmt --all: formats the new probe (the CI fmt failure CodeRabbit
  flagged) and cleans pre-existing drift from #147 (the run_codebase_ingest
  call in main.rs and the #145 tests in rust_ast.rs were merged unformatted).
  Whole workspace is now fmt-clean under the pinned stable rustfmt.

Skipped:

- symbol_hash position-independence: not a defect. A changed file is purged by
  file_key before rewrite (#126 guard), so a skipped pure-line-move leaves the
  vertex _key and start_line mutually consistent and self-healing on the next
  real change. Folding start_line into the digest would force a full re-embed
  on every line shift for no correctness gain.
- DRY-refactor of the test walkers: test-only nitpick. The four small walkers
  are kept separate so the verification test stays obviously correct rather
  than sharing a callback abstraction.

+++ crates/hades-core/src/db/keys.rs
@@ -111,7 +111,9 @@ pub fn symbol_key(file_key: &str, qualified_name: &str, line: usize) -> String {
     let mut hasher = Sha256::new();
     hasher.update(qualified_name.as_bytes());
     hasher.update(b"\n");
-    hasher.update(line.to_le_bytes());
+    // Fixed-width u64 encoding so keys are identical across architectures
+    // (usize::to_le_bytes() is 8 bytes on 64-bit, 4 on 32-bit).
+    hasher.update((line as u64).to_le_bytes());
     let digest = hasher.finalize();
     let hash8 = hex8(&digest);
